% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_prep.R
\name{bt_prep}
\alias{bt_prep}
\title{Prep data function updated for Clarifi Data}
\usage{
bt_prep(
  b,
  align = c("keep.all", "remove.na"),
  dates = NULL,
  fill.gaps = F,
  basic = F
)
}
\arguments{
\item{b}{environment with symbols time series}

\item{align}{alignment type}

\item{dates}{subset of dates}

\item{fill.gaps}{fill gaps introduced by merging}

\item{basic}{control if xts object are created}
}
\value{

}
\description{
Prep data function updated for Clarifi Data
}
