% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quantiles.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{Compute Quantiles updated for Clarifi Data}
\usage{
compute_quantiles(data, next.month.ret, smain = "", n.quantiles = 5, plot = T)
}
\arguments{
\item{data}{factor time series}

\item{next.month.ret}{future returns}

\item{smain}{plot title}

\item{n.quantiles}{number of quantiles}

\item{plot}{flag to create plot}
}
\value{
list with returns, weights, quantiles, ranks
}
\description{
Compute Quantiles updated for Clarifi Data
}
