% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_fund_data.R
\name{date_fund_data}
\alias{date_fund_data}
\title{Date function for fundamental data}
\usage{
date_fund_data(data)
}
\arguments{
\item{data}{fundamental data stored in environment variable}
}
\value{
stored date object
}
\description{
Date function for fundamental data
}
