% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_constituents.R
\name{get_constituents}
\alias{get_constituents}
\title{Get the constituents of a particular quantile for each date in backtest}
\usage{
get_constituents(model_quantile)
}
\arguments{
\item{model_quantile}{single model quantile object}
}
\value{
list of tickers for each date in backtest
}
\description{
Get the constituents of a particular quantile for each date in backtest
}
