% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tiingo_fundamentals.R
\name{merge_tiingo_fundamentals}
\alias{merge_tiingo_fundamentals}
\title{Title}
\usage{
merge_tiingo_fundamentals(data)
}
\arguments{
\item{data}{a data.frame of company financial statements retrieved from `riingo_fundamental_statements()`}
}
\value{
data.frame of combined quarterly fundamentals
}
\description{
Title
}
