% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiingo_prep.R
\name{tiingo_prep}
\alias{tiingo_prep}
\title{Prep fundamental or pricing data from Tiingo API}
\usage{
tiingo_prep(data, type = c("fundamental", "price"))
}
\arguments{
\item{data}{a data.frame of company financial statements retrieved from `riingo_fundamental_statements()` or pricing data from `riingo_prices()`}
}
\value{
data.frame of combined quarterly fundamentals
}
\description{
Prep fundamental or pricing data from Tiingo API
}
